Reorganize challenge category sorting
Currently, the challenges are retrieved from the server and are sorted by the
points. (See /CTFd/api/v1/challenges.py). The categories are sorted by the
maximum point of any challenge within this category. We now change it to
minimum by changing the order of iteration for the loop that creates
categories.